+++ lets/modules/encode/python/compress.py
@@ -67,12 +67,13 @@ class Compress(DockerExtension, Module):
         test = string.ascii_letters + string.digits
         testcmd = "print('%s')" % test
         encoded = b"".join(self.do(testcmd.encode()))
-        cmd = "python -c \"%s\"" % encoded.decode()
+        cmd = encoded.decode()
 
         with self.Container(
             image="python:2",
             network_disabled=True,
-            command=cmd) as container:
+            entrypoint=["python", "-c"],
+            command=[cmd]) as container:
 
             # Fetch output
             output = [line.strip() for line in container.logs(stdout=True, stderr=True)][0]
@@ -86,7 +87,8 @@ class Compress(DockerExtension, Module):
         with self.Container(
             image="python:3",
             network_disabled=True,
-            command=cmd) as container:
+            entrypoint=["python", "-c"],
+            command=[cmd]) as container:
 
             # Fetch output
             output = [line.strip() for line in container.logs(stdout=True, stderr=True)][0]