+++ lets/modules/format/bash/python.py
@@ -69,12 +69,13 @@ class Python(DockerExtension, Module):
         test = string.ascii_letters + string.digits
         testcmd = "echo '%s'" % test
         encoded = b"".join(self.do(testcmd.encode()))
-        cmd = "python -c \"%s\"" % encoded.decode()
+        cmd = encoded.decode()
 
         with self.Container(
             image="python:2",
             network_disabled=True,
-            command=cmd) as container:
+            entrypoint=["python", "-c"],
+            command=[cmd]) as container:
 
             # Fetch output
             output = [line.strip() for line in container.logs(stdout=True, stderr=True)][0]
@@ -88,7 +89,8 @@ class Python(DockerExtension, Module):
         with self.Container(
             image="python:3",
             network_disabled=True,
-            command=cmd) as container:
+            entrypoint=["python", "-c"],
+            command=[cmd]) as container:
 
             # Fetch output
             output = [line.strip() for line in container.logs(stdout=True, stderr=True)][0]